Version 1.4 Pre-alpha 1 (not for use)
⚠️ This build is not for use as basic functionality may not work

- Fixed a bug that caused all default dark themes (assuming dark) in an application to display as light when using a light-colored design theme for Windows applications.
- Started development of functionality related to attaching any windows after selection using “sight”.
Development of the GTK4 version has started (the port is developed using AI, since the author doesn't know a damn thing about GTK). Stable version 1.3 is chosen as the basis, but development is still on Windows with possible porting to Linux (although this is obviously supposed to be different). This version is also Pre-alpha, as it is incredibly unstable and does not contain any working functionality of the original project at all.